msvc/manifest_history: keep msvc-specific code in msvc.rs
The history walker, its three URL/page constants, and the walker unit
test were misplaced in the shared vs_manifest module. They are
MSVC-only — the history mirror serves MSVC manifests and the walker
internally hardcodes find_msvc_candidates(HOST, TARGET) with msvc.rs's
host/target. Moving them into msvc.rs:

- reverts vs_manifest.rs and vs_manifest_tests.rs to identical-to-master
- drops two pub items from the shared module
- lets the walker drop its msvc_ prefix and use the existing HOST/TARGET
  module constants instead of inline string literals
- co-locates the walker test with the function it tests

PR diff is now confined to README + msvc.rs + msvc_tests.rs.
